chore: split out files for _app.page.tsx
